chore : service interface
